+++ NEWS.md
@@ -1,11 +1,15 @@
 # PowerTOST 1.4.9.9999
 
-On GitHub 2020-08-01.
+On GitHub 2020-08-02.
 
 ## Bug fixes
   
   * Fix of the link to package emmeans in man pages of expected power.
   * Fix in Example 1 of `ABE.Rmd` (not a good idea specify a variable with the same name as a function).
+
+## Major changes
+
+  * Pre-compiled RSABE vignette.
 
 ## Minor changes